refactor: extract effectCollection module from calculator (Phase 4a.4)
Move aggregateAttributedEffects and resolveEffectChoices into
calculatorModules/effectCollection.ts. Removes unused imports from
the main calculator. All snapshot tests pass with identical output.
